some import and speed optimizations
git-svn-id: svn://192.168.1.10/public/CoffeeMud@17973 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/WebMacros/PostOfficeBoxNext.java
@@ -80,7 +80,7 @@ public class PostOfficeBoxNext extends StdWebMacro
 			postalBoxes.add(M.Name());
 			for(final Pair<Clan,Integer> C : M.clans())
 			{
-				if(C.first.getAuthority(C.second,Function.WITHDRAW)!=Authority.CAN_NOT_DO)
+				if(C.first.getAuthority(C.second.intValue(),Function.WITHDRAW)!=Authority.CAN_NOT_DO)
 					postalBoxes.add(C.first.name());
 			}
 		}
